Initial adding stacks changes
This includes a lot of edits for clarity, brevity, or removing info
unnecessary for external users to understand new stack introduction.

In addition, most GitHub URLs have been switch to the new monorepo,
avoiding the "this repo is archived" warnings encountered in the old
guide's content.

Still needs some attention to the app service section (I'm considering
moving it to the end), as well as the markup needed for operating in
this website repository (`aside` in particular doesn't have styling).

Finally, I need to double-check the testing section, and rewrite the
earlier portion explain the CoreDB concept within the codebase in a
way that is less complex for outsiders.