Reworded the work_mem section for accuracy. (#495)
The work_mem GUC is particularly tricky because complex queries can allocate a work_mem amount of RAM for every query node in the execution plan. This can result in dozens of instances of work_mem in extreme cases, and can lead to OOM in high connection scenarios. It's one of the harder entries to tune properly, so we should be explicit on the implications and err a bit on the conservative side.