Initial draft for the Plugging the Postgres Upgrade Hole blog post.
This is a (potentially controversial) attempt to address a recent Hacker News
thread centered around the Postgres upgrade process.

https://news.ycombinator.com/item?id=41873957

The overall recommendation is for Postgres to work more like MySQL, and notes
how existing tools may be modified to ease the transition to this end goal.
It also outlines the primarily recommended upgrade procedures so readers have
a complete view of how to upgrade, and the pros/cons for each method, if they
don't already.